requirements.txt
flask
openai
python-dotenv
google-generativeai
transformers
torch
mlx-whisper
